+++ Level0-Module3/src/_03_modulo/_4_amazing_rings/amazing_rings/amazing_rings.pde
@@ -1,0 +1,16 @@
+
+void setup() {
+  
+}
+
+void draw() {
+  
+  // Go to the recipe to run the demonstration before starting this program
+  
+  /*
+  Use the code from your Bullseye program to draw the rings shown in this recipe.
+  You can use the noFill() command to make your ellipses transparent.
+  Make the rings move past each other.
+  When the rings hit the side of the sketch, make them reverse their direction.
+  */
+}
