+++ Level0-Module4/src/_02_boolean/_3_wheres_waldo/wheres_waldo/wheres_waldo.pde
@@ -1,28 +1,41 @@
+/*
 import ddf.minim.*;
 Minim minim = new Minim(this); 
 AudioSample doh;
 AudioSample woohoo;
+*/
 
 void setup() {
+  
+  //Find a Where's Waldo picture and drop it into this sketch.     
   PImage waldo = loadImage("waldo.jpg"); // Change this to match your file name.
-  size(500, 500);
+  size(500, 500); // Change this to match the size of your waldo picture
   image(waldo, 0, 0);
-  doh = minim.loadSample("homer-doh.wav"); //drag and drop from project onto sketch
-  woohoo = minim.loadSample("homer-woohoo.wav"); //drag and drop from project onto sketch } 
+
+  /*
+    doh = minim.loadSample("homer-doh.wav"); 
+    woohoo = minim.loadSample("homer-woohoo.wav"); 
+  */
 }
 
 void draw() {
-      // Use this print statement to find out the coordinates of Waldo
+      // Use this print statement to find out the coordinates of Waldo when you press the mouse
       // println("X: " + mouseX + " Y: " + mouseY); 
 
       // If the mouse is on Waldo, print “Waldo found!”
 
-      // If Waldo is found, also use the method below to play “Woohoo”
+/**********  Now to add the sounds. You will need to uncomment all the code and import the Minim library ********/
+
+      // If Waldo is found, also use a method below to play “Woohoo”
       // Change the name of the sound file if you need to 
+      
       // If the mouse is pressed and they’re not on Waldo, play “Doh”
       // Change the name of the sound file if you need to 
+      
 }
+/*********************  Use the methods below, DON'T CHANGE THE CODE *********************/
 
+/*
 void playWoohoo() {
      woohoo.stop();
      woohoo.trigger();
@@ -32,3 +45,4 @@ void playDoh() {
      doh.stop();
      doh.trigger();
 }
+*/
