+++ Level0-Module4/src/_04_variables_check/_1_Variables_Check.html
@@ -10,7 +10,7 @@
 <body>
 
 <script>
-window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/variableTypesSelfTest.html.html"}
+window.onload=function(){document.location.href="http://central.jointheleague.org/levels/Level0/Recipes/variableTypesSelfTest.html"}
 </script>
